Increase the instance limit to 125
Increasing the instance limit will enable us to experiment with running Spark with lots of tiny worker nodes. 
This doesn't increase the resources the project is allowed to use. The project as a whole will still be pinned to the four physical CascadeLake servers, and limited by the global quotas of 400 cores and 752G RAM.  The current 20 instance limit prevents us from using all 400 cores and 752G RAM because we reach the instance limit before we reach the core or memory limit.